move the scheduling stuff to it's own module
