+++ src/main/python/alppaca/webapp.py
@@ -1,13 +1,9 @@
 from util import init_logging
 from bottle import Bottle
 
-from apscheduler.schedulers.background import BackgroundScheduler
-from apscheduler.triggers.interval import IntervalTrigger
-from apscheduler.events import EVENT_JOB_EXECUTED, EVENT_JOB_ERROR, EVENT_JOB_MISSED
-
-
 from alppaca import IMSInterface
 from alppaca.compat import OrderedDict
+from alppaca.scheduler import configure_scheduler
 
 bottle_app = Bottle(__name__)
 path = '/latest/meta-data/iam/security-credentials/'
@@ -17,8 +13,6 @@ ims_host = 'localhost'
 ims_port = '8080'
 
 credentials = OrderedDict()
-
-task_scheduler = BackgroundScheduler()
 
 logger = init_logging(False)
 
@@ -43,18 +37,6 @@ def refresh_roles():
     credentials = credentials_provider.get_credentials_for_all_roles()
 
 
-def job_executed_event_listener(_):
-    logger.info("Successfully completed credentials refresh")
-
-
-def job_failed_event_listener(event):
-    logger.error("Failed to refresh credentials: {0}".format(event.exception))
-
-
-def job_missed_event_listener(_):
-    logger.warn('Credentials refresh was not executed in time!')
-
-
 def init_roles():
     logger.info("Initializing local credentials cache")
     refresh_roles()
@@ -62,16 +44,10 @@ def init_roles():
 
 def run_scheduler_and_webserver():
     try:
-        task_scheduler.start()
-        trigger = IntervalTrigger(minutes=1)
-        task_scheduler.add_job(func=refresh_roles, trigger=trigger)
-        task_scheduler.add_listener(job_executed_event_listener, EVENT_JOB_EXECUTED)
-        task_scheduler.add_listener(job_failed_event_listener, EVENT_JOB_ERROR)
-        task_scheduler.add_listener(job_missed_event_listener, EVENT_JOB_MISSED)
-
         init_roles()
+        task_scheduler = configure_scheduler(refresh_roles, repeat=1)
+        task_scheduler.start()
         bottle_app.run(host=local_host, port=local_port)
-
     except Exception, e:
         print e
